Move PackageVersionSerializerPlugin into separate file
